moved SearchScreen to stackNavigator, added navbar as a page, moved stackNavigator to app.js
